requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=1.6

Babel>=1.3
cliff>=1.15.0 # Apache-2.0
PyYAML>=3.1.0
